extractor: move to plugins package
